+++ src/spring/HISTORY.md
@@ -5,6 +5,9 @@ Release History
 * Add new command group `az spring application-live-view` to manage Application Live View.
 * Add new command group `az spring dev-tool` to manage Dev Tools.
 * Add argument `--enable-application-live-view` in command `az spring create` to support enable application live view when creating Enterprise sku Spring resource.
+* Add new command group `az spring application-accelerator customized-accelerator` to manage Customized Accelerator.
+* Add new command group `az spring application-accelerator predefined-accelerator` to manage Predefined Accelerator.
+* Add argument `--enable-application-accelerator` in command `az spring create` to support enable application accelerator when creating Enterprise sku Spring resource.
 
 1.2.0
 ---